fix: fix dependency cycle
